# apps/web/openapi.json
{
  "components": {
    "schemas": {
...
      "CoordinatorIn": {
        "properties": {
          "is_coordinator": {
            "title": "Is Coordinator",
            "type": "boolean"
          }
        },
        "required": [
          "is_coordinator"
        ],
        "title": "CoordinatorIn",
        "type": "object"
      },
      "CoordinatorMemberIn": {
        "properties": {
          "coordinator_id": {
            "format": "uuid",
            "title": "Coordinator Id",
            "type": "string"
          },
          "member_id": {
            "format": "uuid",
            "title": "Member Id",
            "type": "string"
          }
        },
        "required": [
          "coordinator_id",
          "member_id"
        ],
        "title": "CoordinatorMemberIn",
        "type": "object"
      },
...
      "DepartmentIn": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "DepartmentIn",
        "type": "object"
      },
      "DepartmentOut": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description"
        ],
        "title": "DepartmentOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API multi-tenant de feedback 360 e avaliação de clientes.",
    "title": "A&S Feedback Hub API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/coordinator-members": {
      "delete": {
        "description": "Remoção de membro audita com `actor_id` e tenta notificar (BR-MIGRAR-026).",
        "operationId": "remove_coordinator_member_api_v1_coordinator_members_delete",
        "parameters": [
          {
            "in": "query",
            "name": "coordinator_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Coordinator Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "member_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Member Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Remove Coordinator Member",
        "tags": [
          "identity-admin"
        ]
      },
      "post": {
        "operationId": "add_coordinator_member_api_v1_coordinator_members_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoordinatorMemberIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Add Coordinator Member Api V1 Coordinator Members Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Coordinator Member",
        "tags": [
          "identity-admin"
        ]
      }
    },
...
    "/api/v1/departments": {
      "get": {
        "operationId": "list_departments_api_v1_departments_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DepartmentOut"
                  },
                  "title": "Response List Departments Api V1 Departments Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Departments",
        "tags": [
          "identity-admin"
        ]
      },
      "post": {
        "operationId": "create_department_api_v1_departments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartmentIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Department",
        "tags": [
          "identity-admin"
        ]
      }
    },
    "/api/v1/departments/{department_id}": {
      "put": {
        "description": "PUT e não PATCH: o corpo é o departamento inteiro, e `description` ausente\nsignifica \"sem descrição\" — não \"mantenha a que estava\".",
        "operationId": "update_department_api_v1_departments__department_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "department_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Department Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartmentIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Department",
        "tags": [
          "identity-admin"
        ]
      }
    },
...
    "/api/v1/profiles": {
      "get": {
        "description": "A tabela de usuários do admin, com o e-mail que mora em `users`.",
        "operationId": "list_profiles_api_v1_profiles_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProfileSummary"
                  },
                  "title": "Response List Profiles Api V1 Profiles Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Profiles",
        "tags": [
          "identity-admin"
        ]
      },
      "post": {
        "description": "Cria credencial e perfil juntos: um sem o outro não entra no sistema.",
        "operationId": "register_user_api_v1_profiles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Register User",
        "tags": [
          "identity-admin"
        ]
      }
    },
    "/api/v1/profiles/{profile_id}": {
      "delete": {
        "description": "Soft-delete: histórico preservado, sessões derrubadas na hora (BR-MIGRAR-018).",
        "operationId": "soft_delete_profile_api_v1_profiles__profile_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Soft Delete Profile",
        "tags": [
          "identity-admin"
        ]
      },
      "patch": {
        "operationId": "update_profile_api_v1_profiles__profile_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Profile",
        "tags": [
          "identity-admin"
        ]
      }
    },
    "/api/v1/profiles/{profile_id}/coordinator": {
      "put": {
        "operationId": "set_coordinator_api_v1_profiles__profile_id__coordinator_put",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoordinatorIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Coordinator",
        "tags": [
          "identity-admin"
        ]
      }
    },
    "/api/v1/profiles/{profile_id}/departments": {
      "put": {
        "operationId": "set_departments_api_v1_profiles__profile_id__departments_put",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartmentsIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Departments",
        "tags": [
          "identity-admin"
        ]
      }
    },
    "/api/v1/profiles/{profile_id}/flags": {
      "patch": {
        "description": "Capacidades individuais (BR-MIGRAR-013/015). Chave desconhecida é 422.",
        "operationId": "set_flags_api_v1_profiles__profile_id__flags_patch",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlagsIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Flags",
        "tags": [
          "identity-admin"
        ]
      }
    },
    "/api/v1/profiles/{profile_id}/manager": {
      "put": {
        "description": "Recusa ciclo na hierarquia — A→B→A faria a resolução de equipe girar sem fim.",
        "operationId": "assign_manager_api_v1_profiles__profile_id__manager_put",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManagerIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Assign Manager",
        "tags": [
          "identity-admin"
        ]
      }
    },
    "/api/v1/profiles/{profile_id}/password": {
      "post": {
        "description": "Redefinir senha derruba as sessões abertas: se a conta estava comprometida, o\ntoken do invasor continuaria valendo.",
        "operationId": "reset_password_api_v1_profiles__profile_id__password_post",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reset Password",
        "tags": [
          "identity-admin"
        ]
      }
    },
    "/api/v1/profiles/{profile_id}/reactivate": {
      "post": {
        "operationId": "reactivate_profile_api_v1_profiles__profile_id__reactivate_post",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reactivate Profile",
        "tags": [
          "identity-admin"
        ]
      }
    },
    "/api/v1/profiles/{profile_id}/role": {
      "put": {
        "operationId": "change_role_api_v1_profiles__profile_id__role_put",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Change Role",
        "tags": [
          "identity-admin"
        ]
      }
    },
...
  }
}